Seeds filters

+++ libs/seedCarOwners.js
@@ -4,7 +4,10 @@ function seedCarOwners(carOwners) {
   CarOwners.find()
     .countDocuments()
     .then((totalCarOwners) => {
-      if (totalCarOwners < carOwners.length) {
+      if (
+        totalCarOwners < carOwners.length &&
+        process.env.NODE_ENV !== "development"
+      ) {
         console.log("Seeding Car Owners");
         CarOwners.deleteMany({}).exec();
         for (let i = 0; i < carOwners.length; i++) {
